Refactored data consistency and Authentication logic

+++ backend/src/main/java/com/oracle/football/dto/PlayerReportDto.java
@@ -3,7 +3,7 @@ package com.oracle.football.dto;
 import java.util.List;
 
 public record PlayerReportDto(
-        Long reportId,
+        Integer reportId,
         String playerName,
         String playerStrengths,
         String playerWeaknesses,
